requirements.txt
# My Green Doctor - Dependencies
# ================================
# This file lists all Python packages required to run the application.
# Install with: pip install -r requirements.txt

# Web Framework
streamlit>=1.28.0

# Google Gemini AI SDK
google-generativeai>=0.3.0

# Image Processing
Pillow>=10.0.0

